fix: Strip whitespace from input
While furiously typing input (while not understanding the timing
mechanism) there was a tendency to add extra whitespace.  As it doesn't
seem to be working against the spirit of the gameplay I added a strip()
function to each of the input calls to further normalize the input
beyond just rendering lowercase.

Signed-Off: Brian 'redbeard' Harrington <redbeard@dead-city.org>